automatic npmjs deploy
travis will automatically deploy the built package on npmjs when a commit
in the master branch is tagged. the npm package version is taken from the
version in Cargo.toml. the tag name doesn't matter, but it's best if it
follows the package version (perhaps prefixed with v - for example
v0.1.0).

the `edge: true` option enables the use of dpl v2, since the newer npm
provider has the `src:` option.

https://blog.travis-ci.com/2019-08-27-deployment-tooling-dpl-v2-preview-release